+++ typings/rx-document.d.ts
@@ -47,7 +47,7 @@ export declare class RxDocumentBase<RxDocumentType, OrmMethods = {}> {
     readonly allAttachments$: Observable<RxAttachment<RxDocumentType, OrmMethods>[]>;
 
     toJSON(): RxDocumentTypeWithRev<RxDocumentType>;
-    toJSON(withRev: false): RxDocumentType;
+    toJSON(withRevAndAttachments: false): RxDocumentType;
 
     destroy(): void;
 }